requirements.txt
Flask==2.3.3
python-docx==0.8.11
PyPDF2==3.0.1
spacy==3.6.1
python-dotenv==1.0.0
Werkzeug==2.3.7 